Support for multiple languages added. This only works for the website so far. Offline docs are still English-only.
Non-English files have the extension .<lang>.php, eg. compiling.it.php
is the Italian version of the English file compiling.php. If there's no
translation of a particular file, the page for that language will have
English content, but the translated template will be used, and internal
links will be to other pages in the same language.

We still need to add links to the various translations from the English 
home page.

Also, some missing Subversion properties have been added.


git-svn-id: https://svn.aros.org/svn/aros/trunk@24462 fb15a70f-31f2-0310-bbcc-cdcc74a49acc